requirements.txt
langchain
faiss-cpu
sentence-transformers
requests
python-dotenv
unstructured
playwright
pymongo
Flask
gunicorn

